+++ src/Engine.h
@@ -3,6 +3,6 @@
 
 void Engine_Start(void);
 
-void Engine_Run(const decimal deltaTime);
+void Engine_Run(const double deltaTime);
 
 void Engine_Shutdown(void);
